Moving things around - adding ticker handler
